move extractVideFrames to videIndexer class
