+++ app/src/main/java/com/fpf/smartscan/lib/VideoUtils.kt
@@ -7,44 +7,9 @@ import android.graphics.Bitmap
 import android.media.MediaMetadataRetriever
 import android.net.Uri
 import android.provider.MediaStore
-import android.util.Log
 import androidx.core.graphics.scale
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-
-fun extractFramesFromVideo(context: Context, videoUri: Uri, frameCount: Int = 10): List<Bitmap>? {
-    val retriever = MediaMetadataRetriever()
-    return try {
-        retriever.setDataSource(context, videoUri)
-
-        val durationUs = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION)?.toLong()?.times(1000)
-            ?: return null
-
-        val frameList = mutableListOf<Bitmap>()
-
-        for (i in 0 until frameCount) {
-            val frameTimeUs = (i * durationUs) / frameCount
-            val bitmap = retriever.getFrameAtTime(frameTimeUs, MediaMetadataRetriever.OPTION_CLOSEST_SYNC)
-
-            if (bitmap != null) {
-                frameList.add(bitmap)
-            } else {
-                // Temporary Fix: Break early if null which suggest codec issue with video
-                break
-            }
-        }
-
-        if (frameList.isEmpty()) return null
-
-        frameList
-    } catch (e: Exception) {
-        Log.e("FrameExtractionError", "Error extracting video frames: $e")
-        null
-    } finally {
-        retriever.release()
-    }
-}
-
 
 fun getVideoUriFromId(id: Long): Uri {
     return ContentUris.withAppendedId(
